backend/requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.1.13
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.3.1
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-vision==3.5.0
google-currency==1.0.10
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.6
inflection==0.5.1
Pillow==10.1.0
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
python-dotenv==0.21.0
pytz==2023.3.post1
requests==2.31.0
rest-framework-simplejwt==0.0.2
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
